Allow the user to overwrite configuration files
We use the term 'overwrite', but of course what the script does is to
remove the exiting .rc file before attempting to create the symlink in
the home directory. If the user enters an unrecognized command it is
assumed to mean 'ignore' — that is, we do not do anything by default.