Enable dynamic windows titles
The shelltitle command makes it possible that while we are running a
command the title of the window automatically changes to that command.
Note that this involves modifying .bashrc so that our shell outputs a
null title-escape-sequence as part of the prompt, but that file is not
yet ready to be version controlled and shown to the world.