Use Rafael Magana's hardstatus line
The main difference with respect to our previous hardstatus line is
that this one includes the current time and date. Also, the list of
windows is padded to the display's width — i.e., they are centered.
I have modified the command so that red background is used to mark
the active window instead of parentheses.

Kudos to Rafael Magana for making his screenrc available on GitHub:
https://github.com/rafmagana/screenrc