Revert "Redefine command key to C-z"
The idea of rebinding the command key of screen to C-z, as we did in
commit 4e968a6fbc7a9a24b49441, was not as good as it looked at first.
An unexpected consequence of getting used to this key binding is that
I keep suspending the current foreground job when working on machines
where my screenrc is not installed, which is really annoying. I must
resign myself to use `C-a a` to send the command character (C-a) to
Emacs — this is indeed mentioned in the man page for screen!